# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SDN/AAC Internal Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /crews:
    get:
      operationId: listCrews
      tags: [crews]
      parameters:
        - name: seasonId
          in: query
          schema:
            type: integer
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Crews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Crew"
    post:
      operationId: createCrew
      tags: [crews]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CrewInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Crew"
  /crews/{id}:
    get:
      operationId: getCrew
      tags: [crews]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Crew
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Crew"
    patch:
      operationId: updateCrew
      tags: [crews]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CrewUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Crew"
    delete:
      operationId: deleteCrew
      tags: [crews]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Fleet ─────────────────────────────────────────────────────────────────────
...
  /equipment:
    get:
      operationId: listEquipment
      tags: [equipment]
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Equipment list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Equipment"
    post:
      operationId: createEquipment
      tags: [equipment]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EquipmentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Equipment"
  /equipment/{id}:
    get:
      operationId: getEquipment
      tags: [equipment]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Equipment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Equipment"
    patch:
      operationId: updateEquipment
      tags: [equipment]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EquipmentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Equipment"
    delete:
      operationId: deleteEquipment
      tags: [equipment]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Training Schedules ────────────────────────────────────────────────────────
...
  /documents:
    get:
      operationId: listDocuments
      tags: [documents]
      parameters:
        - name: type
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
    post:
      operationId: createDocument
      tags: [documents]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
  /documents/{id}:
    get:
      operationId: getDocument
      tags: [documents]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
    patch:
      operationId: updateDocument
      tags: [documents]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
    delete:
      operationId: deleteDocument
      tags: [documents]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Dashboard ─────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Athlete:
      type: object
      required: [id, name, birthDate, gender, status, memberNumber, affiliationDate]
      properties:
        id:
          type: integer
        name:
          type: string
        birthDate:
          type: string
          format: date
        gender:
          type: string
          enum: [M, F]
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        memberNumber:
          type: ["string", "null"]
        fprNumber:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        affiliationDate:
          type: string
          format: date
        status:
          type: string
          enum: [ativo, inativo, suspenso]
        notes:
          type: ["string", "null"]
        categoryOverride:
          type: ["string", "null"]
...
    Crew:
      type: object
      required: [id, name, boatClass, category, seasonId]
      properties:
        id:
          type: integer
        name:
          type: string
        boatClass:
          type: string
          enum: ["1x", "2x", "2-", "4x", "4-", "4+", "8+"]
        category:
          type: string
        seasonId:
          type: integer
        seasonName:
          type: string
        athleteIds:
          type: array
          items:
            type: integer
        athletes:
          type: array
          items:
            $ref: "#/components/schemas/Athlete"
    CrewInput:
      type: object
      required: [name, boatClass, category, seasonId]
      properties:
        name:
          type: string
        boatClass:
          type: string
          enum: ["1x", "2x", "2-", "4x", "4-", "4+", "8+"]
        category:
          type: string
        seasonId:
          type: integer
        athleteIds:
          type: array
          items:
            type: integer
    CrewUpdate:
      type: object
      properties:
        name:
          type: string
        boatClass:
          type: string
          enum: ["1x", "2x", "2-", "4x", "4-", "4+", "8+"]
        category:
          type: string
        seasonId:
          type: integer
        athleteIds:
          type: array
          items:
            type: integer

    # ── Fleet ───────────────────────────────────────────────────────────────────
...
    Equipment:
      type: object
      required: [id, name, category, totalQuantity, availableQuantity, status]
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
        totalQuantity:
          type: integer
        availableQuantity:
          type: integer
        status:
          type: string
        assignedTo:
          type: ["string", "null"]
        acquisitionDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
    EquipmentInput:
      type: object
      required: [name, category, totalQuantity, availableQuantity, status]
      properties:
        name:
          type: string
        category:
          type: string
        totalQuantity:
          type: integer
        availableQuantity:
          type: integer
        status:
          type: string
        assignedTo:
          type: ["string", "null"]
        acquisitionDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
    EquipmentUpdate:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        totalQuantity:
          type: integer
        availableQuantity:
          type: integer
        status:
          type: string
        assignedTo:
          type: ["string", "null"]
        acquisitionDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]

    # ── Training Schedules ──────────────────────────────────────────────────────
...
    Document:
      type: object
      required: [id, title, type, date]
      properties:
        id:
          type: integer
        title:
          type: string
        type:
          type: string
          enum: [noticia, contrato, arquivo]
        date:
          type: string
          format: date
        content:
          type: ["string", "null"]
        fileUrl:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        entity:
          type: ["string", "null"]
        contractStart:
          type: ["string", "null"]
        contractEnd:
          type: ["string", "null"]
        contractStatus:
          type: ["string", "null"]
          enum: [ativo, expirado, null]
        notes:
          type: ["string", "null"]
    DocumentInput:
      type: object
      required: [title, type, date]
      properties:
        title:
          type: string
        type:
          type: string
          enum: [noticia, contrato, arquivo]
        date:
          type: string
          format: date
        content:
          type: ["string", "null"]
        fileUrl:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        entity:
          type: ["string", "null"]
        contractStart:
          type: ["string", "null"]
        contractEnd:
          type: ["string", "null"]
        contractStatus:
          type: ["string", "null"]
          enum: [ativo, expirado, null]
        notes:
          type: ["string", "null"]
    DocumentUpdate:
      type: object
      properties:
        title:
          type: string
        type:
          type: string
          enum: [noticia, contrato, arquivo]
        date:
          type: string
          format: date
        content:
          type: ["string", "null"]
        fileUrl:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        entity:
          type: ["string", "null"]
        contractStart:
          type: ["string", "null"]
        contractEnd:
          type: ["string", "null"]
        contractStatus:
          type: ["string", "null"]
          enum: [ativo, expirado, null]
        notes:
          type: ["string", "null"]

    # ── Dashboard ────────────────────────────────────────────────────────────────
...