# lib/api-zod/src/generated/types/crew.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * SDN/AAC Internal Management API
 * OpenAPI spec version: 0.1.0
 */
import type { Athlete } from './athlete';
import type { CrewBoatClass } from './crewBoatClass';

export interface Crew {
  id: number;
  name: string;
  boatClass: CrewBoatClass;
  category: string;
  seasonId: number;
  seasonName?: string;
  athleteIds?: number[];
  athletes?: Athlete[];
}